updata

+++ utilcode/src/main/java/com/blankj/utilcode/util/CrashUtils.java
@@ -1,5 +1,6 @@
 package com.blankj.utilcode.util;
 
+import android.annotation.SuppressLint;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.os.Build;
@@ -14,7 +15,6 @@ import java.lang.Thread.UncaughtExceptionHandler;
 import java.text.Format;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.Locale;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
@@ -36,16 +36,21 @@ public final class CrashUtils {
     private static ExecutorService sExecutor;
 
     private static final String FILE_SEP = System.getProperty("file.separator");
-    private static final Format FORMAT   = new SimpleDateFormat("MM-dd HH-mm-ss", Locale.getDefault());
+    @SuppressLint("SimpleDateFormat")
+    private static final Format FORMAT   = new SimpleDateFormat("MM-dd HH-mm-ss");
 
     private static final String CRASH_HEAD;
 
     private static final UncaughtExceptionHandler DEFAULT_UNCAUGHT_EXCEPTION_HANDLER;
     private static final UncaughtExceptionHandler UNCAUGHT_EXCEPTION_HANDLER;
 
+    private static OnCrashListener sOnCrashListener;
+
     static {
         try {
-            PackageInfo pi = Utils.getApp().getPackageManager().getPackageInfo(Utils.getApp().getPackageName(), 0);
+            PackageInfo pi = Utils.getApp()
+                    .getPackageManager()
+                    .getPackageInfo(Utils.getApp().getPackageName(), 0);
             if (pi != null) {
                 versionName = pi.versionName;
                 versionCode = pi.versionCode;
@@ -69,9 +74,16 @@ public final class CrashUtils {
             @Override
             public void uncaughtException(final Thread t, final Throwable e) {
                 if (e == null) {
-                    android.os.Process.killProcess(android.os.Process.myPid());
-                    System.exit(0);
+                    if (DEFAULT_UNCAUGHT_EXCEPTION_HANDLER != null) {
+                        DEFAULT_UNCAUGHT_EXCEPTION_HANDLER.uncaughtException(t, null);
+                    } else {
+                        android.os.Process.killProcess(android.os.Process.myPid());
+                        System.exit(1);
+                    }
                     return;
+                }
+                if (sOnCrashListener != null) {
+                    sOnCrashListener.onCrash(e);
                 }
                 Date now = new Date(System.currentTimeMillis());
                 String fileName = FORMAT.format(now) + ".txt";
@@ -115,7 +127,7 @@ public final class CrashUtils {
 
     /**
      * 初始化
-     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>}</p>
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
      */
     public static void init() {
         init("");
@@ -123,21 +135,53 @@ public final class CrashUtils {
 
     /**
      * 初始化
-     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>}</p>
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
      *
      * @param crashDir 崩溃文件存储目录
      */
     public static void init(@NonNull final File crashDir) {
-        init(crashDir.getAbsolutePath());
+        init(crashDir.getAbsolutePath(), null);
     }
 
     /**
      * 初始化
-     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>}</p>
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
      *
      * @param crashDir 崩溃文件存储目录
      */
     public static void init(final String crashDir) {
+        init(crashDir, null);
+    }
+
+    /**
+     * 初始化
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
+     *
+     * @param onCrashListener 崩溃监听事件
+     */
+    public static void init(final OnCrashListener onCrashListener) {
+        init("", onCrashListener);
+    }
+
+    /**
+     * 初始化
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
+     *
+     * @param crashDir        崩溃文件存储目录
+     * @param onCrashListener 崩溃监听事件
+     */
+    public static void init(@NonNull final File crashDir, final OnCrashListener onCrashListener) {
+        init(crashDir.getAbsolutePath(), onCrashListener);
+    }
+
+    /**
+     * 初始化
+     * <p>需添加权限 {@code <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />}</p>
+     *
+     * @param crashDir        崩溃文件存储目录
+     * @param onCrashListener 崩溃监听事件
+     */
+    public static void init(final String crashDir, final OnCrashListener onCrashListener) {
         if (isSpace(crashDir)) {
             dir = null;
         } else {
@@ -149,6 +193,7 @@ public final class CrashUtils {
         else {
             defaultDir = Utils.getApp().getCacheDir() + FILE_SEP + "crash" + FILE_SEP;
         }
+        sOnCrashListener = onCrashListener;
         Thread.setDefaultUncaughtExceptionHandler(UNCAUGHT_EXCEPTION_HANDLER);
     }
 
@@ -176,5 +221,9 @@ public final class CrashUtils {
             }
         }
         return true;
+    }
+
+    public interface OnCrashListener {
+        void onCrash(Throwable e);
     }
 }
